// src/app/models/user.ts
/* 
 * Copyright 2019 Ignacio Loyola @nodonade.com
 * Version 0.1 (Working first step blog)
 */

export interface User {
    // id: number,
    // name: string,
    // surname: string,
    // role: string,
    // email: string,
    // password: string,
    // description: string,
    // image: string
   uid: string;
   email: string;
   displayName: string;
   photoURL: string;
   emailVerified: boolean;
   // bio: string;
}